Billboard/Data-Collection/requirements.txt
"""
Python dependencies for Billboard Hot 100 data collection
"""

# Core dependencies
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Database
psycopg2-binary>=2.9.9

# Google Trends
pytrends>=4.9.2

# Environment variables
python-dotenv>=1.0.0

# Data processing
pandas>=2.1.0
numpy>=1.24.0

# Scheduling (for automated collection)
schedule>=1.2.0
APScheduler>=3.10.4

# Logging
python-logging>=0.4.9.6

# Date/time utilities
python-dateutil>=2.8.2

